+++ components/exam-main/manage-exam.module.scss
@@ -136,6 +136,152 @@
     0 0 10px rgba(136, 151, 108, 0.1); // Cập nhật màu sắc
 }
 
+.btnColor1 {
+  gap: 0.5rem;
+  border: none;
+  transition: all 0.4s cubic-bezier(0.165, 0.84, 0.44, 1);
+  border-radius: 10px;
+  font-weight: 600;
+  place-content: center;
+  font-size: 0.9rem;
+  line-height: 1rem;
+  height: 38px;
+  font-size: 12px;
+
+  background-color: #46DFB1; // Màu mới
+  color: #fff;
+  box-shadow: 0 4px 6px rgba(70, 223, 177, 0.2), 0 0 10px rgba(70, 223, 177, 0.1);
+}
+
+.btnColor1:hover {
+  background-color: #39BFA3; // Màu mới khi hover
+  box-shadow: 0 6px 15px rgba(57, 191, 163, 0.3), 0 0 30px rgba(57, 191, 163, 0.2),
+    0 0 60px rgba(57, 191, 163, 0.1);
+  transform: translateY(-2px);
+}
+
+.btnColor1:active {
+  transform: translateY(1px);
+  box-shadow: 0 2px 4px rgba(57, 191, 163, 0.2), 0 0 10px rgba(57, 191, 163, 0.1);
+}
+
+.btnColor2 {
+  gap: 0.5rem;
+  border: none;
+  transition: all 0.4s cubic-bezier(0.165, 0.84, 0.44, 1);
+  border-radius: 10px;
+  font-weight: 600;
+  place-content: center;
+  font-size: 0.9rem;
+  line-height: 1rem;
+  height: 38px;
+  font-size: 12px;
+
+  background-color: #09D1C7; // Màu mới
+  color: #fff;
+  box-shadow: 0 4px 6px rgba(9, 209, 199, 0.2), 0 0 10px rgba(9, 209, 199, 0.1);
+}
+
+.btnColor2:hover {
+  background-color: #07B8B0; // Màu mới khi hover
+  box-shadow: 0 6px 15px rgba(7, 184, 176, 0.3), 0 0 30px rgba(7, 184, 176, 0.2),
+    0 0 60px rgba(7, 184, 176, 0.1);
+  transform: translateY(-2px);
+}
+
+.btnColor2:active {
+  transform: translateY(1px);
+  box-shadow: 0 2px 4px rgba(7, 184, 176, 0.2), 0 0 10px rgba(7, 184, 176, 0.1);
+}
+
+.btnColor3 {
+  gap: 0.5rem;
+  border: none;
+  transition: all 0.4s cubic-bezier(0.165, 0.84, 0.44, 1);
+  border-radius: 10px;
+  font-weight: 600;
+  place-content: center;
+  font-size: 0.9rem;
+  line-height: 1rem;
+  height: 38px;
+  font-size: 12px;
+  
+  background-color: #15919B; // Màu mới
+  color: #fff;
+  box-shadow: 0 4px 6px rgba(21, 145, 155, 0.2), 0 0 10px rgba(21, 145, 155, 0.1);
+}
+
+.btnColor3:hover {
+  background-color: #0E7A7F; // Màu mới khi hover
+  box-shadow: 0 6px 15px rgba(14, 122, 127, 0.3), 0 0 30px rgba(14, 122, 127, 0.2),
+    0 0 60px rgba(14, 122, 127, 0.1);
+  transform: translateY(-2px);
+}
+
+.btnColor3:active {
+  transform: translateY(1px);
+  box-shadow: 0 2px 4px rgba(14, 122, 127, 0.2), 0 0 10px rgba(14, 122, 127, 0.1);
+}
+
+.btnColor4 {
+  gap: 0.5rem;
+  border: none;
+  transition: all 0.4s cubic-bezier(0.165, 0.84, 0.44, 1);
+  border-radius: 10px;
+  font-weight: 600;
+  place-content: center;
+  font-size: 0.9rem;
+  line-height: 1rem;
+  height: 38px;
+  font-size: 12px;
+
+  background-color: #0C6478; // Màu mới
+  color: #fff;
+  box-shadow: 0 4px 6px rgba(12, 100, 120, 0.2), 0 0 10px rgba(12, 100, 120, 0.1);
+}
+
+.btnColor4:hover {
+  background-color: #0A4D5B; // Màu mới khi hover
+  box-shadow: 0 6px 15px rgba(10, 77, 91, 0.3), 0 0 30px rgba(10, 77, 91, 0.2),
+    0 0 60px rgba(10, 77, 91, 0.1);
+  transform: translateY(-2px);
+}
+
+.btnColor4:active {
+  transform: translateY(1px);
+  box-shadow: 0 2px 4px rgba(10, 77, 91, 0.2), 0 0 10px rgba(10, 77, 91, 0.1);
+}
+
+.btnColor5 {
+  gap: 0.5rem;
+  border: none;
+  transition: all 0.4s cubic-bezier(0.165, 0.84, 0.44, 1);
+  border-radius: 10px;
+  font-weight: 600;
+  place-content: center;
+  font-size: 0.9rem;
+  line-height: 1rem;
+  height: 38px;
+  font-size: 12px;
+
+  background-color: #213A58; // Màu mới
+  color: #fff;
+  box-shadow: 0 4px 6px rgba(33, 58, 88, 0.2), 0 0 10px rgba(33, 58, 88, 0.1);
+}
+
+.btnColor5:hover {
+  background-color: #1A2A3E; // Màu mới khi hover
+  box-shadow: 0 6px 15px rgba(26, 42, 62, 0.3), 0 0 30px rgba(26, 42, 62, 0.2),
+    0 0 60px rgba(26, 42, 62, 0.1);
+  transform: translateY(-2px);
+}
+
+.btnColor5:active {
+  transform: translateY(1px);
+  box-shadow: 0 2px 4px rgba(26, 42, 62, 0.2), 0 0 10px rgba(26, 42, 62, 0.1);
+}
+
+
 .btnSmall {
   width: 100px !important; // Loại bỏ dấu ngoặc kép
   height: 38px !important; // Loại bỏ dấu ngoặc kép
@@ -166,6 +312,13 @@
   font-size: 12px;
 }
 
+.btnLarge2 {
+  width: 300px !important; // Loại bỏ dấu ngoặc kép
+  height: 38px !important; // Loại bỏ dấu ngoặc kép
+  font-size: 12px;
+}
+
+
 .textFieldCustom {
   & fieldset {
     border-color: green;
@@ -188,4 +341,8 @@
   font-weight: bold; /* Đậm chữ */
   background-color: #0e4e79; /* Màu nền khi được chọn */
   color: white; /* Màu chữ khi được chọn */
+}
+
+.btnBorderRadius {
+  border-radius: 10px; /* Bo tròn viền */
 }
