+++ pages/exam/list-answer-by-student.tsx
@@ -1,0 +1,20 @@
+import { GetServerSideProps } from "next";
+import { getSession } from "next-auth/react";
+import Head from "next/head";
+import { useRef } from "react";
+import LoadingBar, { LoadingBarRef } from "react-top-loading-bar";
+import NavBarHome from "../../components/home/navbar-home";
+import ListAnswerByStudent from "../../components/exam-main/list-answer-by-student";
+
+const ListAnwByStudent= () => {
+  const loadingBarRef: React.Ref<LoadingBarRef> = useRef(null);
+
+  return (
+    <div>
+      <NavBarHome loadingBarRef={loadingBarRef} />
+      <ListAnswerByStudent />
+    </div>
+  );
+};
+
+export default ListAnwByStudent;
