+++ pages/exam/list-grade.tsx
@@ -1,0 +1,20 @@
+import { GetServerSideProps } from "next";
+import { getSession } from "next-auth/react";
+import Head from "next/head";
+import { useRef } from "react";
+import LoadingBar, { LoadingBarRef } from "react-top-loading-bar";
+import NavBarHome from "../../components/home/navbar-home";
+import ListGradeForm from "../../components/exam-main/list-grade";
+
+const ListGrade = () => {
+  const loadingBarRef: React.Ref<LoadingBarRef> = useRef(null);
+
+  return (
+    <div>
+      <NavBarHome loadingBarRef={loadingBarRef} />
+      <ListGradeForm />
+    </div>
+  );
+};
+
+export default ListGrade;
