--- pages/exam/view-exam.tsx
@@ -1,20 +1,0 @@
-import { GetServerSideProps } from "next";
-import { getSession } from "next-auth/react";
-import Head from "next/head";
-import { useRef } from "react";
-import LoadingBar, { LoadingBarRef } from "react-top-loading-bar";
-import NavBarHome from "../../components/home/navbar-home";
-import ViewExamForm from "../../components/exam-main/view-exam";
-
-const ViewExam = () => {
-    const loadingBarRef: React.Ref<LoadingBarRef> = useRef(null);
-
-    return (
-        <div>
-            <NavBarHome loadingBarRef={loadingBarRef} />
-            <ViewExamForm />
-        </div>
-    );
-};
-
-export default ViewExam;
